Move all waveforms to single file and adjust pulse package and tests
