+++ qctoolkit/pulses/conditions.py
@@ -14,7 +14,7 @@ from typing import Dict, Optional, Callable
 from qctoolkit.utils.types import ChannelID
 from qctoolkit.pulses.parameters import Parameter
 from qctoolkit.pulses.sequencing import SequencingElement, Sequencer
-from qctoolkit.pulses.instructions import InstructionBlock, InstructionPointer, Trigger
+from qctoolkit._program.instructions import InstructionBlock, InstructionPointer, Trigger
 
 __all__ = ["Condition", "ConditionEvaluationException", "ConditionMissingException",
            "SoftwareCondition", "HardwareCondition"]
