+++ qctoolkit/pulses/function_pulse_template.py
@@ -19,10 +19,10 @@ from qctoolkit.serialization import Serializer, PulseRegistryType
 from qctoolkit.utils.types import ChannelID, TimeType, time_from_float
 from qctoolkit.pulses.parameters import Parameter, ParameterConstrainer, ParameterConstraint
 from qctoolkit.pulses.pulse_template import AtomicPulseTemplate, MeasurementDeclaration
-from qctoolkit.pulses.instructions import Waveform
+from qctoolkit._program.waveforms import FunctionWaveform
 
 
-__all__ = ["FunctionPulseTemplate", "FunctionWaveform"]
+__all__ = ["FunctionPulseTemplate"]
 
 
 class FunctionPulseTemplate(AtomicPulseTemplate, ParameterConstrainer):
@@ -161,49 +161,3 @@ class FunctionPulseTemplate(AtomicPulseTemplate, ParameterConstrainer):
         )}
 
 
-class FunctionWaveform(Waveform):
-    """Waveform obtained from instantiating a FunctionPulseTemplate."""
-
-    def __init__(self, expression: ExpressionScalar,
-                 duration: float,
-                 channel: ChannelID) -> None:
-        """Creates a new FunctionWaveform instance.
-
-        Args:
-            expression: The function represented by this FunctionWaveform
-                as a mathematical expression where 't' denotes the time variable. It must not have other variables
-            duration: The duration of the waveform
-            measurement_windows: A list of measurement windows
-            channel: The channel this waveform is played on
-        """
-        super().__init__()
-        if set(expression.variables) - set('t'):
-            raise ValueError('FunctionWaveforms may not depend on anything but "t"')
-
-        self._expression = expression
-        self._duration = time_from_float(duration)
-        self._channel_id = channel
-
-    @property
-    def defined_channels(self) -> Set[ChannelID]:
-        return {self._channel_id}
-    
-    @property
-    def compare_key(self) -> Any:
-        return self._channel_id, self._expression, self._duration
-
-    @property
-    def duration(self) -> TimeType:
-        return self._duration
-
-    def unsafe_sample(self,
-                      channel: ChannelID,
-                      sample_times: np.ndarray,
-                      output_array: Union[np.ndarray, None] = None) -> np.ndarray:
-        if output_array is None:
-            output_array = np.empty(len(sample_times))
-        output_array[:] = self._expression.evaluate_numeric(t=sample_times)
-        return output_array
-
-    def unsafe_get_subset_for_channels(self, channels: Set[ChannelID]) -> Waveform:
-        return self
