+++ qctoolkit/pulses/multi_channel_pulse_template.py
@@ -17,120 +17,14 @@ import numpy
 from qctoolkit.serialization import Serializer, PulseRegistryType
 
 from qctoolkit.utils.types import ChannelID, TimeType
-from qctoolkit.pulses.instructions import Waveform
+from qctoolkit._program.waveforms import MultiChannelWaveform
 from qctoolkit.pulses.pulse_template import PulseTemplate, AtomicPulseTemplate
 from qctoolkit.pulses.pulse_template_parameter_mapping import MappingPulseTemplate, MappingTuple
 from qctoolkit.pulses.parameters import Parameter, ParameterConstrainer
 from qctoolkit.pulses.measurement import MeasurementDeclaration
 from qctoolkit.expressions import Expression, ExpressionScalar
 
-__all__ = ["MultiChannelWaveform", "AtomicMultiChannelPulseTemplate"]
-
-
-class MultiChannelWaveform(Waveform):
-    """A MultiChannelWaveform is a Waveform object that allows combining arbitrary Waveform objects
-    to into a single waveform defined for several channels.
-
-    The number of channels used by the MultiChannelWaveform object is the sum of the channels used
-    by the Waveform objects it consists of.
-
-    MultiChannelWaveform allows an arbitrary mapping of channels defined by the Waveforms it
-    consists of and the channels it defines. For example, if the MultiChannelWaveform consists
-    of a two Waveform objects A and B which define two channels each, then the channels of the
-    MultiChannelWaveform may be 0: A.1, 1: B.0, 2: B.1, 3: A.0 where A.0 means channel 0 of Waveform
-    object A.
-
-    The following constraints must hold:
-     - The durations of all Waveform objects must be equal.
-     - The channel mapping must be sane, i.e., no channel of the MultiChannelWaveform must be
-        assigned more than one channel of any Waveform object it consists of
-    """
-
-    def __init__(self, sub_waveforms: Iterable[Waveform]) -> None:
-        """Create a new MultiChannelWaveform instance.
-
-        Requires a list of subwaveforms in the form (Waveform, List(int)) where the list defines
-        the channel mapping, i.e., a value y at index x in the list means that channel x of the
-        subwaveform will be mapped to channel y of this MultiChannelWaveform object.
-
-        Args:
-            sub_waveforms (Iterable( Waveform )): The list of sub waveforms of this
-                MultiChannelWaveform
-        Raises:
-            ValueError, if a channel mapping is out of bounds of the channels defined by this
-                MultiChannelWaveform
-            ValueError, if several subwaveform channels are assigned to a single channel of this
-                MultiChannelWaveform
-            ValueError, if subwaveforms have inconsistent durations
-        """
-        super().__init__()
-        if not sub_waveforms:
-            raise ValueError(
-                "MultiChannelWaveform cannot be constructed without channel waveforms."
-            )
-
-        # avoid unnecessary multi channel nesting
-        def flatten_sub_waveforms(to_flatten):
-            for sub_waveform in to_flatten:
-                if isinstance(sub_waveform, MultiChannelWaveform):
-                    yield from sub_waveform._sub_waveforms
-                else:
-                    yield sub_waveform
-
-        # sort the waveforms with their defined channels to make compare key reproducible
-        def get_sub_waveform_sort_key(waveform):
-            return tuple(sorted(tuple('{}_stringified_numeric_channel'.format(ch) if isinstance(ch, int) else ch
-                                      for ch in waveform.defined_channels)))
-
-        self._sub_waveforms = sorted(flatten_sub_waveforms(sub_waveforms),
-                                     key=get_sub_waveform_sort_key)
-
-        if not all(waveform.duration == self._sub_waveforms[0].duration for waveform in self._sub_waveforms[1:]):
-            raise ValueError(
-                "MultiChannelWaveform cannot be constructed from channel waveforms of different"
-                "lengths."
-            )
-        self.__defined_channels = set()
-        for waveform in self._sub_waveforms:
-            if waveform.defined_channels & self.__defined_channels:
-                raise ValueError('Channel may not be defined in multiple waveforms')
-            self.__defined_channels |= waveform.defined_channels
-
-    @property
-    def duration(self) -> TimeType:
-        return self._sub_waveforms[0].duration
-
-    def __getitem__(self, key: ChannelID) -> Waveform:
-        for waveform in self._sub_waveforms:
-            if key in waveform.defined_channels:
-                return waveform
-        raise KeyError('Unknown channel ID: {}'.format(key), key)
-
-    @property
-    def defined_channels(self) -> Set[ChannelID]:
-        return self.__defined_channels
-
-    @property
-    def compare_key(self) -> Any:
-        # sort with channels
-        return tuple(sub_waveform.compare_key for sub_waveform in self._sub_waveforms)
-
-    def unsafe_sample(self,
-                      channel: ChannelID,
-                      sample_times: numpy.ndarray,
-                      output_array: Union[numpy.ndarray, None]=None) -> numpy.ndarray:
-        return self[channel].unsafe_sample(channel, sample_times, output_array)
-
-    def unsafe_get_subset_for_channels(self, channels: Set[ChannelID]) -> 'Waveform':
-        relevant_sub_waveforms = tuple(swf for swf in self._sub_waveforms if swf.defined_channels & channels)
-        if len(relevant_sub_waveforms) == 1:
-            return relevant_sub_waveforms[0].get_subset_for_channels(channels)
-        elif len(relevant_sub_waveforms) > 1:
-            return MultiChannelWaveform(
-                sub_waveform.get_subset_for_channels(channels & sub_waveform.defined_channels)
-                for sub_waveform in relevant_sub_waveforms)
-        else:
-            raise KeyError('Unknown channels: {}'.format(channels))
+__all__ = ["AtomicMultiChannelPulseTemplate"]
 
 
 class AtomicMultiChannelPulseTemplate(AtomicPulseTemplate, ParameterConstrainer):
