+++ qctoolkit/pulses/pulse_template.py
@@ -18,7 +18,7 @@ from qctoolkit.expressions import ExpressionScalar
 from qctoolkit.pulses.conditions import Condition
 from qctoolkit.pulses.parameters import Parameter
 from qctoolkit.pulses.sequencing import Sequencer, SequencingElement, InstructionBlock
-from qctoolkit.pulses.instructions import Waveform
+from qctoolkit._program.waveforms import Waveform
 from qctoolkit.pulses.measurement import MeasurementDefiner, MeasurementDeclaration
 
 __all__ = ["PulseTemplate", "AtomicPulseTemplate", "DoubleParameterNameException", "MappingTuple"]
