+++ tests/pulses/conditions_tests.py
@@ -1,7 +1,7 @@
 import unittest
 from typing import Optional
 
-from qctoolkit.pulses.instructions import InstructionPointer, Trigger, CJMPInstruction, GOTOInstruction
+from qctoolkit._program.instructions import InstructionPointer, Trigger, CJMPInstruction, GOTOInstruction
 from qctoolkit.pulses.conditions import HardwareCondition, SoftwareCondition, ConditionEvaluationException
 
 from tests.pulses.sequencing_dummies import DummySequencingElement, DummySequencer, DummyInstructionBlock, DummyInstruction
