+++ tests/pulses/measurement_tests.py
@@ -4,7 +4,7 @@ from qctoolkit.pulses.parameters import ParameterConstraint, ParameterConstraint
     ParameterNotProvidedException, ParameterConstrainer, ConstantParameter
 from qctoolkit.pulses.measurement import MeasurementDefiner
 
-from qctoolkit.pulses.instructions import InstructionBlock, MEASInstruction
+from qctoolkit._program.instructions import InstructionBlock, MEASInstruction
 
 
 class MeasurementDefinerTest(unittest.TestCase):
