+++ tests/pulses/sequencing_tests.py
@@ -1,7 +1,7 @@
 import unittest
 
 from qctoolkit.pulses.parameters import  ConstantParameter
-from qctoolkit.pulses.instructions import InstructionBlock, STOPInstruction
+from qctoolkit._program.instructions import InstructionBlock, STOPInstruction
 from qctoolkit.pulses.sequencing import Sequencer
 
 from tests.pulses.sequencing_dummies import DummySequencingElement, DummyCondition
